Edited the database. I also created a Seeder and Factory so that the dummy can be placed
