+++ app/Http/Controllers/TransactionController.php
@@ -4,12 +4,17 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
-use App\Http\Controllers\TransactionController;
+use App\Models\Transaction;
 
 class TransactionController extends Controller
 {
     public function index()
     {
-        return view('transaction.transaction');
+        // すべてのトランザクションデータを取得
+        $transactions = Transaction::all();
+
+        return view('transaction.index', compact('transactions'));
     }
 }
+
+
