+++ app/Models/Cart.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace App\Models;
+
+use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Database\Eloquent\Model;
+
+class Cart extends Model
+{
+    use HasFactory;
+    protected $table = 'carts';
+
+    public function user()
+    {
+        return $this->belongsTo(User::class, 'cart_id');
+    }
+}
