+++ app/Models/Category.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace App\Models;
+
+use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Database\Eloquent\Model;
+
+class Category extends Model
+{
+    use HasFactory;
+    protected $table = 'categories';
+
+    public function user()
+    {
+        return $this->belongsTo(User::class, 'category_id');
+    }
+}
