+++ app/Models/Item.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace App\Models;
+
+use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Database\Eloquent\Model;
+
+class Item extends Model
+{
+    use HasFactory;
+    protected $table = 'items';
+
+    public function user()
+    {
+        return $this->belongsTo(User::class, 'item_id');
+    }
+}
