+++ app/Models/Role.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace App\Models;
+
+use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Database\Eloquent\Model;
+
+class Role extends Model
+{
+    use HasFactory;
+    protected $table = 'roles';
+
+    public function user()
+    {
+        return $this->belongsTo(User::class, 'role_id');
+    }
+}
