+++ database/seeders/TransactionSeeder.php
@@ -1,0 +1,21 @@
+<?php
+
+namespace Database\Seeders;
+
+use App\Models\Transaction;
+use App\Models\User;
+use Illuminate\Database\Seeder;
+
+class TransactionSeeder extends Seeder
+{
+    /**
+     * Run the database seeds.
+     *
+     * @return void
+     */
+    public function run()
+    {
+        
+        Transaction::factory()->count(10)->create();
+    }
+}
