+++ resources/sass/styles.scss
@@ -1,5 +1,5 @@
 /* Apply Poppins font to body text */
-@import "~font-awesome/css/font-awesome.css";
+// @import "~font-awesome/css/font-awesome.css";
 
 @import url('https://cdn-uicons.flaticon.com/uicons-solid-rounded/css/uicons-solid-rounded.css');
