+++ resources/views/layouts/header.blade.php
@@ -12,22 +12,22 @@
           <div class="collapse navbar-collapse" id="navbarNav">
             <ul class="navbar-nav">
               <li class="nav-item">
-                <a class="nav-link" href="{{ route('home') }}"><i class="fi fi-sr-home" style="margin-right: 5px;"></i>HOME</a>
+                <a class="nav-link" href="#"><i class="fi fi-sr-home" style="margin-right: 5px;"></i>HOME</a>
               </li>
               <li class="nav-item">
-                <a class="nav-link" href="#"><i class="fi fi-sr-layout-fluid" style="margin-right: 5px;"></i>Item List</a>
+                <a class="nav-link" href="{{ route('home') }}"><i class="fi fi-sr-layout-fluid" style="margin-right: 5px;"></i>Item List</a>
               </li>
               <li class="nav-item">
-                <a class="nav-link" href="item.edit"><i class="fi fi-sr-edit" style="margin-right: 5px;"></i>Edite item</a>
+                <a class="nav-link" href="{{ route('category.edit')}}"><i class="fi fi-sr-edit" style="margin-right: 5px;"></i>Edite item</a>
               </li>
               <li class="nav-item">
-                <a class="nav-link" href="category.edit"><i class="fi fi-sr-home" style="margin-right: 5px;"></i>Categories</a>
+                <a class="nav-link" href="#"><i class="fi fi-sr-home" style="margin-right: 5px;"></i>Categories</a>
               </li>
               <li class="nav-item">
-                <a class="nav-link" href="#"><i class="fi fi-sr-chart-histogram" style="margin-right: 5px;"></i>Dashboard</a>
+                <a class="nav-link" href="{{ route('home') }}"><i class="fi fi-sr-chart-histogram" style="margin-right: 5px;"></i>Dashboard</a>
               </li>
               <li class="nav-item">
-                <a class="nav-link" href="#"><i class="fi-sr-users-alt" style="margin-right: 5px;"></i>Casher register</a>
+                <a class="nav-link" href="{{ route('userlist.index') }}"><i class="fi-sr-users-alt" style="margin-right: 5px;"></i>Casher register</a>
               </li>
               <li class="nav-item">
                 <a class="nav-link" href="{{ route('transaction.index') }}"><i class="fi fi-sr-ballot" style="margin-right: 5px;"></i>Transaction</a>
