+++ webpack.mix.js
@@ -18,4 +18,6 @@ mix.js('resources/js/app.js', 'public/js')
     .sass('resources/sass/login-forget.scss', 'public/css')
     .sass('resources/sass/register.scss', 'public/css')
     .sass('resources/sass/styles.scss', 'public/css')
+    .sass('resources/sass/dashboard.scss', 'public/css')
+    .js('resources/js/dashboard.js', 'public/js')
     .sourceMaps();
